Adding order tests
